requirements.txt
fastapi[all]
Authlib==1.2.0
google-auth==2.15.0
itsdangerous==2.1.2
pydantic==1.10.2
SQLAlchemy==1.4.45
starlette==0.22.0
uvicorn==0.20.0
requests==2.28.1
httpx==0.23.1
openpyxl
pandas
google-oauth
oauth